Refactor name from banding data fields to maps reference data
further refactoring of dictionary needed
